add common typing package `@docusaurus/typing-common` .
