+++ .eslintignore
@@ -14,3 +14,4 @@ packages/docusaurus-utils/lib/
 packages/docusaurus/lib/
 packages/docusaurus-init/lib/
 packages/docusaurus-plugin-content-blog/lib/
+packages/docusaurus-typing-common/lib/
