+++ .gitignore
@@ -16,3 +16,4 @@ packages/docusaurus-utils/lib/
 packages/docusaurus/lib/
 packages/docusaurus-init/lib/
 packages/docusaurus-plugin-content-blog/lib/
+packages/docusaurus-typing-common/lib/
