+++ packages/docusaurus-plugin-content-blog/package.json
@@ -12,6 +12,7 @@
   "license": "MIT",
   "dependencies": {
     "@docusaurus/mdx-loader": "^2.0.0-alpha.24",
+    "@docusaurus/typing-common": "^2.0.0-alpha.24",
     "@docusaurus/utils": "^2.0.0-alpha.24",
     "fs-extra": "^8.1.0",
     "globby": "^10.0.1",
