+++ packages/docusaurus-plugin-content-blog/src/index.ts
@@ -19,13 +19,13 @@ import {
   BlogContent,
   BlogItemsToModules,
   TagsModule,
-  ConfigureWebpackUtils,
 } from './types';
 import {
   LoadContext,
   PluginContentLoadedActions,
   RouteModule,
-} from './typesDocusaurus';
+  ConfigureWebpackUtils,
+} from '@docusaurus/typing-common';
 import {Configuration} from 'webpack';
 
 // YYYY-MM-DD-{name}.mdx?
@@ -57,7 +57,7 @@ export default function pluginContentBlog(
   context: LoadContext,
   opts: Partial<PluginOptions>,
 ) {
-  const options = {...DEFAULT_OPTIONS, ...opts};
+  const options: PluginOptions = {...DEFAULT_OPTIONS, ...opts};
   const contentPath = path.resolve(context.siteDir, options.path);
 
   return {
